apps/ml-service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
redis==5.0.1
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.1.post1
pydantic==2.6.1
python-dotenv==1.0.1
websockets==12.0
httpx==0.26.0
xgboost==2.0.3
tensorflow==2.15.0
